requirements.txt
pyrogram==2.0.106  # Для работы с Telegram API
python-dotenv==1.0.0  # Для работы с переменными окружения
redislite==6.0.415  # Легковесная Redis-база данных

